+++ packager/react-packager/src/AssetServer/index.js
@@ -13,7 +13,7 @@ const Promise = require('promise');
 const crypto = require('crypto');
 const declareOpts = require('../lib/declareOpts');
 const fs = require('fs');
-const getAssetDataFromName = require('../lib/getAssetDataFromName');
+const getAssetDataFromName = require('../DependencyResolver/lib/getAssetDataFromName');
 const path = require('path');
 
 const stat = Promise.denodeify(fs.stat);
