+++ packager/react-packager/src/DependencyResolver/AssetModule.js
@@ -2,7 +2,7 @@
 
 const Module = require('./Module');
 const Promise = require('promise');
-const getAssetDataFromName = require('../lib/getAssetDataFromName');
+const getAssetDataFromName = require('./lib/getAssetDataFromName');
 
 class AssetModule extends Module {
   constructor(...args) {
