+++ packager/react-packager/src/DependencyResolver/DependencyGraph/ResolutionRequest.js
@@ -12,7 +12,7 @@ const debug = require('debug')('ReactNativePackager:DependencyGraph');
 const util = require('util');
 const path = require('path');
 const isAbsolutePath = require('absolute-path');
-const getAssetDataFromName = require('../../lib/getAssetDataFromName');
+const getAssetDataFromName = require('../lib/getAssetDataFromName');
 const Promise = require('promise');
 
 class ResolutionRequest {
