+++ packager/react-packager/src/DependencyResolver/lib/__tests__/getAssetDataFromName-test.js
