+++ packager/react-packager/src/DependencyResolver/lib/__tests__/getPlatformExtension-test.js
