+++ packager/react-packager/src/DependencyResolver/lib/getAssetDataFromName.js
