+++ packager/react-packager/src/DependencyResolver/lib/getPlatformExtension.js
